Refactor token getter
