+++ src/lib/email.ts
@@ -1,4 +1,37 @@
-import { getAccessToken } from "./googleAuth";
+let cachedToken: string | null = null;
+let tokenExpiry = 0;
+
+export async function getAccessToken() {
+  const now = Date.now();
+
+  if (cachedToken && now < tokenExpiry) {
+    return cachedToken;
+  }
+
+  const params = new URLSearchParams({
+    client_id: process.env.GOOGLE_CLIENT_ID!,
+    client_secret: process.env.GOOGLE_CLIENT_SECRET!,
+    refresh_token: process.env.GMAIL_REFRESH_TOKEN!,
+    grant_type: "refresh_token",
+  });
+
+  const response = await fetch("https://oauth2.googleapis.com/token", {
+    method: "POST",
+    headers: { "Content-Type": "application/x-www-form-urlencoded" },
+    body: params.toString(),
+  });
+
+  if (!response.ok) {
+    throw new Error("Failed to refresh access token");
+  }
+
+  const data = await response.json();
+
+  cachedToken = data.access_token;
+  tokenExpiry = now + data.expires_in * 1000 - 60000;
+
+  return cachedToken;
+}
 
 export const sendConfirmationEmail = async (recipient: string, subject: string) => {
   const accessToken = await getAccessToken();
